requirements.txt
fastapi==0.115.12
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.20.1
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow==11.2.1
pydantic==2.11.4
pydantic_settings==2.9.1
python-dotenv==1.1.0
Requests==2.32.3
scikit-image
streamlit==1.45.1
